Move Processor Behavior to a Module
